requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
pydantic==1.10.7
python-dotenv==1.0.0
pytz==2023.3
pywebio==1.8.2
SQLAlchemy==2.0.9
tornado==6.3.1
typing_extensions==4.5.0
ua-parser==0.16.1
user-agents==2.2.0
Werkzeug==2.2.3
yarl==1.8.2
